+++ src/Symfony/Component/VarDumper/CHANGELOG.md
@@ -5,6 +5,8 @@ CHANGELOG
 -----
 
  * added `AbstractCloner::setMinDepth()` function to ensure minimum tree depth
+ * added `HtmlDumper::markHeaderAsDumped()` function to prevent dumping HTML header
+ * increased visibility of `HtmlDumper::getDumpHeader()` function from protected to public
 
 2.7.0
 -----